Initial integration of NVMe, replacing OSD in exofs
This patch begins the transition of exofs from using OSD devices to NVMe
devices for mounting and file operations.
Modifies kernel configuration to build NVMe core and block driver
as built-in (`y` instead of `m`). Adds a new function in nvme module
to submit key-value commands to an NVMe device. Updates exofs functions
to read and write using NVMe commands.
In additions, add TODOs for further refactoring and investigation of
NVMe integration issues.

Co-authored-by: Amit Shoham <amitshoham8950@gmail.com>, Adam Zats <adamrnd2@gmail.com>
Change-Id: Ib533da31bfc0c503ea60d64d9ba0fc022246d295